KOGITO-2697 Rename File to ElectronFile
