+++ packages/desktop/src/webview/home/FilesPage.tsx
@@ -45,7 +45,7 @@ import { SortAlphaDownIcon } from "@patternfly/react-icons";
 import * as electron from "electron";
 import * as React from "react";
 import { useCallback, useContext, useEffect, useMemo, useState } from "react";
-import { ElectronFile, UNSAVED_FILE_NAME } from "../../common/File";
+import { ElectronFile, UNSAVED_FILE_NAME } from "../../common/ElectronFile";
 import { RecentOpenedFile } from "../../common/RecentOpenedFile";
 import { extractFileExtension, removeDirectories } from "../../common/utils";
 import { GlobalContext } from "../common/GlobalContext";
